+++ src/vznncv/cubemx/tools/_project_description.py
@@ -1,6 +1,7 @@
-from collections import namedtuple
-
+import os
 import re
+from collections import namedtuple
+from os.path import abspath, splitext, relpath, join, dirname
 
 
 def parse_make_var_array(var_value):
@@ -68,8 +69,12 @@ ProjectDescription = namedtuple('ProjectDescription', [
     'ld_script'
 ])
 
+_HEADER_FILE_EXTENSIONS = ['.h', '.hpp']
+_SRC_FILE_EXTENSIONS = ['.c', '.cpp', '.s']
 
-def build_project_description(make_vars, project_dir, optimization_flags=None):
+
+def build_project_description(make_vars, project_dir, optimization_flags=None, autodiscover_sources=False):
+    project_dir = abspath(project_dir)
     build_dir = parse_make_var_path(make_vars.get('BUILD_DIR', 'build'))
     target = parse_make_var_single_value(make_vars['TARGET'])
 
@@ -77,13 +82,30 @@ def build_project_description(make_vars, project_dir, optimization_flags=None):
     for var_name, var_val in make_vars.items():
         if var_name.lower().endswith('_sources'):
             source_files.update(parse_make_var_paths(var_val))
-    source_files = sorted(source_files)
 
     include_dirs = set()
     for var_name, var_val in make_vars.items():
         if var_name.lower().endswith('_includes'):
             include_dirs.update(parse_make_var_includes(var_val))
-    include_dirs = sorted(include_dirs)
+
+    if autodiscover_sources:
+        for root, _, files in os.walk(project_dir):
+            base_dir = relpath(root, project_dir)
+            has_header = False
+            for filename in files:
+                file_ext = splitext(filename)[1]
+                if file_ext in _HEADER_FILE_EXTENSIONS:
+                    has_header = True
+                elif file_ext in _SRC_FILE_EXTENSIONS:
+                    source_files.add(join(base_dir, filename))
+            if has_header:
+                root = abspath(root)
+                while root != project_dir:
+                    include_dirs.add(relpath(root, project_dir))
+                    parent_root = dirname(root)
+                    if parent_root == root:
+                        break
+                    root = parent_root
 
     c_defs = parse_make_var_defines(make_vars['C_DEFS'])
     as_defs = parse_make_var_defines(make_vars['AS_DEFS'])
@@ -100,8 +122,8 @@ def build_project_description(make_vars, project_dir, optimization_flags=None):
         stm_series=stm_series,
         project_dir=project_dir,
 
-        source_files=source_files,
-        include_dirs=include_dirs,
+        source_files=sorted(source_files),
+        include_dirs=sorted(include_dirs),
         definitions=sorted(set(c_defs) | set(as_defs)),
         mcu_flags=mcu_flags,
         optimization_flags=optimization_flags or '',